The-Zookeeper/requirements.txt
disnake==2.4.0
python-dotenv==0.19.2
aiohttp==3.8.1